requirements.txt
pyyaml
openai
pandas